+++ test/src/dos_loadseg.c
@@ -6,13 +6,13 @@ int main(int argc, char *argv[])
   BPTR seglist;
 
   if(argc != 2) {
-    Printf("Usage: %s <file>\n", argv[0]);
+    Printf("Usage: %s <file>\n", (ULONG)argv[0]);
     return 1;
   }
 
   seglist = LoadSeg(argv[1]);
   if(seglist == 0) {
-    Printf("No seglist found: %s\n", argv[1]);
+    Printf("No seglist found: %s\n", (ULONG)argv[1]);
     return 2;
   }
